backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
google-adk>=1.0.0
google-genai>=1.0.0
google-cloud-firestore>=2.19.0
python-dotenv==1.0.1
pydantic==2.9.2
python-multipart==0.0.12
passlib==1.7.4
bcrypt==3.2.2
